+++ src/main/java/shared/Convert.java
@@ -1,0 +1,26 @@
+package shared;
+
+public class UnitUtils {
+    
+    /**
+     * Converts meters to feet.
+     *
+     * @param      meters  The meters quantity to convert
+     *
+     * @return     feet
+     */
+    public static double metersToFeet(double meters) {
+        return meters * 3.28084;
+    }
+
+    /**
+     * Converts feet to meters.
+     *
+     * @param      feet  The feet quantity to convert
+     *
+     * @return     meters
+     */
+    public static double feetToMeters(double feet) {
+        return feet / 3.28084;
+    }
+}
